Add routes to modal pages and create mirror SEO pages
